Refactor: prepare for interactive mode
- New synchronous mode forces tracer to synchronise with sanitizer after each message;
- User output is now implemented by delegate outside of the sanitizer;
- The sanitizer now remember the entire process tree.
- The sanitizer now stores cmdlines of processes and updates them on exec syscalls;
- Changed structure/field names to better represent their roles:
  - TracerConnectionData -> TracerProcess;
  - MakeConnectionData -> MakeConnectionHandler;
  - TargetDatabase -> MakeProcess;
- tracer_stub executable can be used instead of the real tracer to pass eavesdropper output back to the sanitizer;
- Fixed protocol message type name collision (between INODE_UNLINK and INIT);
- PID reuse problem was addressed and fixed with PidEpoch structure;
- Reworked structure responsibilities;
- Added --fix option to ./check-format.sh;
- Add formatting rules for .sh scripts;

+++ check-format.sh
@@ -1,6 +1,21 @@
 #!/bin/bash
 
-if ! git diff --quiet; then
+fix_flag=false
+
+while [ $# -gt 0 ]; do
+    case $1 in
+    -f | --fix)
+        fix_flag=true
+        ;;
+    *)
+        echo "Usage: $(basename "$0") [-f|--fix]"
+        exit 1
+        ;;
+    esac
+    shift
+done
+
+if ! $fix_flag && ! git diff --quiet; then
     echo >&2 "There are uncommitted changes in the working directory."
     echo >&2 "Please commit or stash them before running this script."
     exit 1
@@ -27,7 +42,7 @@ formatting_errors=0
 
 clang-format -i "${files[@]}"
 
-if ! git diff --quiet; then
+if ! $fix_flag && ! git diff --quiet; then
     echo "The following files are not formatted correctly:"
     git diff --name-only
     git reset --hard HEAD
